requirements-dev.txt
-i https://pypi.org/simple
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
click==7.0
coverage==4.5.2
dependency-injector==3.12.4
flask==1.0.2
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
jinja2==2.10
lazy-object-proxy==1.3.1
locustio==0.9.0
markupsafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
msgpack==0.6.0
pep8==1.7.1
pluggy==0.8.0
ptvsd==4.2.0
py==1.7.0
pycodestyle==2.4.0
pylint-django==2.0.5
pylint-plugin-utils==0.4
pylint==2.2.2
pytest-cov==2.6.1
pytest-django==3.4.5
pytest==4.1.0
pyzmq==17.1.2
requests==2.21.0
six==1.11.0
urllib3==1.24.1
werkzeug==0.14.1
wrapt==1.10.11
yapf==0.25.0
